Move multi-key map to utils
